Bring all the changes coming from the de-id project
