+++ {{cookiecutter.project_dash_slug}}/setup.cfg
@@ -31,7 +31,24 @@ install_requires =
   {%- if cookiecutter.command_line_interface|lower == 'click' %}
   newrelic
   Click>=6.0
+  python-decouple
+  bump2version
+  newrelic
+  pytest-cov
+  envparse
   {%- endif %}
+
+[options.extras_require]
+# These are optional dependencies that are not needed for simply running the code.
+  dev =
+    awscli-local
+    flake8
+  test =
+    pytest-cov
+    unittest-xml-reporting
+    pytest-ordering
+  prod =
+    newrelic
 
 [mypy]
 ignore_missing_imports = true
