--- {{cookiecutter.project_dash_slug}}/{{cookiecutter.project_slug}}/cli.py
@@ -1,18 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Console script for {{ cookiecutter.project_name }}."""
-import sys
-import click
-
-
-@click.command()
-def main(args=None):
-    """Console script for {{cookiecutter.project_slug}}."""
-    click.echo("Replace this message by putting your code into "
-               "{{cookiecutter.project_slug}}.cli.main")
-    click.echo("See click documentation at http://click.pocoo.org/")
-    return 0
-
-
-if __name__ == "__main__":
-    sys.exit(main())  # pragma: no cover
