+++ {{cookiecutter.project_dash_slug}}/{{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}.py
@@ -1,5 +1,19 @@
 # -*- coding: utf-8 -*-
 
-"""Main module."""
+"""Console script for {{ cookiecutter.project_name }}."""
+import sys
+import click
+from decouple import config
 
-from settings import Config
+
+@click.command()
+def main(args=None):
+    """Console script for {{cookiecutter.project_slug}}."""
+    click.echo("Replace this message by putting your code into "
+               "{{cookiecutter.project_slug}}.cli.main")
+    click.echo("See click documentation at http://click.pocoo.org/")
+    return 0
+
+
+if __name__ == "__main__":
+    sys.exit(main())  # pragma: no cover
